Move tests out of `ui/UtilsTest.kt` to match changes in `main/`
